css/css-values/getComputedStyle-calc-background-size-percent-016.html

<!DOCTYPE html>

  <meta charset="UTF-8">

  <title>CSS Values Test: computed value of background-size: calc(67% - 50% + 4em)</title>


  <!--
  "
  Where percentages are not resolved at computed-value time,
  they are not resolved in math functions, e.g.
  'calc(100% - 100% + 1px)' resolves to 'calc(0% + 1px)', not to
  '1px'. If there are special rules for computing percentages
  in a value (e.g. the 'height' property), they apply whenever
  a math function contains percentages.
  "
  § 10.11 Computed Value
  https://www.w3.org/TR/css-values-4/#calc-computed-value
  -->

  <link rel="author" title="Gérard Talbot" href="http://www.gtalbot.org/BrowserBugsSection/css21testsuite/">
  <link rel="help" href="https://www.w3.org/TR/css-values-4/#calc-computed-value">

  <meta name="flags" content="">
  <meta content="This test verifies that a term with a percentage unit that can not be resolved at computed-value time will require a calc() wrapper. A term with an em value, on the other hand, can be resolved at computed-value time and therefore must be absolutized to 'px'." name="assert">

  <script src="/resources/testharness.js"></script>

  <script src="/resources/testharnessreport.js"></script>

  <style>
  html, body
    {
      font-size: 16px;
      height: 100%;
      margin: 0px 8px;
    }

  div#target
    {
      background-color: yellow;
      background-image: url("support/cat.png");
      background-position: top center;
      background-repeat: no-repeat;
      height: 140px;
    }
  </style>

  <div id="target"></div>

  <script>
  function startTesting()
  {

  var targetElement = document.getElementById("target");

    function verifyComputedStyle(property_name, specified_value, expected_value, description)
    {

    test(function()
      {

      targetElement.style.setProperty(property_name, specified_value);

      assert_equals(getComputedStyle(targetElement)[property_name], expected_value);

      }, description);
    }

    verifyComputedStyle("background-size", "calc(67% - 50% + 4em)", "calc(17% + 64px) auto", "testing background-size: calc(67% - 50% + 4em)");

    /*
    "Where percentages are not resolved at computed-value time,
    they are not resolved in math functions (...)"
    https://www.w3.org/TR/css-values-4/#calc-serialize

    Therefore here, the percentage is preserved and
    a calc() wrapper must be used. The 4em term
    must be resolved though.
    */

  }

  startTesting();

  </script>
